generate sw-manifest in prod, set resource version for replacement
