+++ package.json
@@ -2,7 +2,7 @@
   "name": "web.dev",
   "version": "2.0.0",
   "scripts": {
-    "build": "npm-run-all clean sass gulp rollup eleventy rollup-sw",
+    "build": "npm-run-all clean sass gulp rollup eleventy sw-manifest",
     "clean": "rm -rf dist",
     "coverage:integration": "c8 npm-run-all test:lib test:unit test:integration",
     "coverage:unit": "c8 npm run test:unit",
@@ -14,7 +14,7 @@
     "debug:test:unit": "node --inspect-brk ./node_modules/.bin/mocha ./test/unit",
     "deploy:cleanup": "node ./tools/cleanup-versions/index.js web-dev-production-1",
     "deploy": "ELEVENTY_ENV=prod npm run build && node index-algolia.js && gcloud app deploy --project web-dev-production-1 --quiet",
-    "dev": "npm-run-all clean rollup rollup-sw sass --parallel --race watch:rollup watch:sass watch:eleventy watch:gulp watch:rollup-sw start",
+    "dev": "npm-run-all clean rollup sass --parallel --race watch:rollup watch:sass watch:eleventy watch:gulp start",
     "eleventy": "eleventy --quiet",
     "gulp": "npx gulp build",
     "karma": "karma start",
@@ -27,13 +27,13 @@
     "updated": "node ./tools/update-updated/index.js",
     "translated": "node ./tools/update-translated/index.js",
     "rollup": "node build.js",
-    "rollup-sw": "node build-sw.js",
     "sass": "node ./compile-css.js",
     "snapshots": "percy exec --config ./tools/percy/.percy.conf.yml -- node tools/percy/snapshots.js",
     "stage:cleanup": "node ./tools/cleanup-versions/index.js web-dev-staging",
     "stage:personal": "ELEVENTY_ENV=prod npm run build && gcloud app deploy --project web-dev-staging --quiet --no-promote",
     "stage": "ELEVENTY_ENV=prod npm run build && gcloud app deploy --project web-dev-staging --quiet",
     "start": "node ./server.js",
+    "sw-manifest": "node sw-manifest.js",
     "test:integration": "mocha ./test/integration",
     "test:lib": "npm-run-all clean rollup karma",
     "test:unit": "mocha ./test/unit && npm run test:lib",
@@ -41,7 +41,6 @@
     "watch:eleventy": "eleventy --watch --quiet",
     "watch:gulp": "npx gulp build --silent && gulp watch --silent",
     "watch:rollup": "chokidar \"src/lib/**/*.{js,scss}\" --silent -c \"npm run rollup\"",
-    "watch:rollup-sw": "chokidar \"src/lib/**/*.js\" --silent -c \"npm run rollup-sw\"",
     "watch:sass": "chokidar \"src/styles/**/*.scss\" --silent -c \"npm run sass\"",
     "watch:test:lib": "TEST_MODE=dev npm-run-all clean rollup --parallel watch:rollup karma"
   },
