+++ src/site/_data/.gitignore
@@ -1,0 +1,3 @@
+# Temporary files generated by builds. This is here as Eleventy won't see
+# .gitignore files in subdirectories.
+resource*.json
